new Cylinder structure
