+++ magpylib/_lib/fields/field_wrap_BH_v.py
@@ -66,7 +66,7 @@ def getBHv_level2(**kwargs: dict) -> np.ndarray:
             dim = np.array(kwargs['dimension'], dtype=float)
             tile_params['dimension'] = (dim,2)
 
-        elif src_type == 'Cylinder_old':
+        elif src_type == 'CylinderSection':
             mag = np.array(kwargs['magnetization'], dtype=float)
             tile_params['magnetization'] = (mag,2)
             dim = np.array(kwargs['dimension'], dtype=float)
@@ -95,6 +95,10 @@ def getBHv_level2(**kwargs: dict) -> np.ndarray:
             tile_params['segment_start'] = (pos_start,2)
             pos_end = np.array(kwargs['segment_end'], dtype=float)
             tile_params['segment_end'] = (pos_end,2)
+
+        else:
+            msg = 'Unknown source_type'
+            raise MagpylibBadUserInput(msg)
 
     except KeyError as kerr:
         msg = f'Missing input keys: {str(kerr)}'
