+++ magpylib/magnet/__init__.py
@@ -4,6 +4,6 @@ magnetization given in units of millitesla [mT] through mu0*M. See documentation
 for details on field computations.
 """
 
-__all__ = ['Cuboid', 'Cylinder', 'Sphere']
+__all__ = ['Cuboid', 'Cylinder', 'Sphere', 'CylinderSection']
 
-from magpylib._lib.obj_classes import Cuboid, Cylinder, Sphere
+from magpylib._lib.obj_classes import Cuboid, Cylinder, Sphere, CylinderSection
