+++ tests/test_obj_Cylinder.py
@@ -34,6 +34,12 @@ def test_repr():
     pm2 = magpy.magnet.Cylinder((1,2,3),(2,3))
     assert pm2.__repr__()[:8] == 'Cylinder', 'Cylinder repr failed'
 
+def test_repr2():
+    """ test __repr__
+    """
+    pm2 = magpy.magnet.CylinderSection((1,2,3),(2,3,1,0,45))
+    assert pm2.__repr__()[:15] == 'CylinderSection', 'CylinderSection repr failed'
+
 
 def test_Cylinder_getBH():
     """
@@ -43,68 +49,45 @@ def test_Cylinder_getBH():
     mag = (22,33,44)
     poso = (np.random.rand(100, 3)-.5)*5
 
-    dim2 = (1,2)
-    dim2_5 = (1,2,0,0,360)
+    dim2 = [(1,2), (2,3), (3,4)]
+    dim5 = [(0,1,2,0,360), (0,2,3,0,360), (0,3,4,0,360)]
 
-    dim3 = (1,2,.5)
-    dim3_5 = (1,2,.5,0,360)
+    for d2,d5 in zip(dim2,dim5):
 
-    dim5 = (2.1,5,1.2,30,145)
+        src1 = magpy.magnet.Cylinder(mag, d2)
+        src2 = magpy.magnet.CylinderSection(mag, d5)
+        B0 = src1.getB(poso)
+        H0 = src1.getH(poso)
 
-    for dim,dim6 in zip([dim2, dim3, dim5], [dim2_5, dim3_5, dim5]):
-        src = magpy.magnet.Cylinder(mag, dim)
-        B1 = src.getB(poso)
-        H1 = src.getH(poso)
+        B1 = src2.getB(poso)
+        H1 = src2.getH(poso)
 
         B2 = magpy.getBv(
             source_type='Cylinder',
             magnetization=mag,
-            dimension=dim,
+            dimension=d2,
             observer=poso)
         H2 = magpy.getHv(
             source_type='Cylinder',
             magnetization=mag,
-            dimension=dim,
+            dimension=d2,
             observer=poso)
 
         B3 = magpy.getBv(
-            source_type='Cylinder',
+            source_type='CylinderSection',
             magnetization=mag,
-            dimension=dim6,
+            dimension=d5,
             observer=poso)
         H3 = magpy.getHv(
-            source_type='Cylinder',
+            source_type='CylinderSection',
             magnetization=mag,
-            dimension=dim6,
+            dimension=d5,
             observer=poso)
 
         assert np.allclose(B1, B2)
         assert np.allclose(B1, B3)
+        assert np.allclose(B1, B0)
 
         assert np.allclose(H1, H2)
         assert np.allclose(H1, H3)
-
-
-def test_Cylinder_vs_old_inside():
-    """
-    test Cylinder vs old version
-    """
-    # inside
-    mag = (np.random.rand(10,3)-.5)*1000
-    dim = np.random.rand(10,2)+1
-    poso = (np.random.rand(10, 3)-.5)
-
-    magpy.Config.ITER_CYLINDER = 1000
-    H_new = magpy.getHv(
-        source_type='Cylinder',
-        magnetization=mag,
-        dimension=dim,
-        observer=poso)
-    H_old = magpy.getHv(
-        source_type='Cylinder_old',
-        magnetization=mag,
-        dimension=dim,
-        observer=poso)
-
-    err = np.linalg.norm(H_new-H_old)/np.linalg.norm(H_new)
-    assert err<1e-1
+        assert np.allclose(H1, H0)
